PriceFetcher/pyproject.toml
[project]
name = "pricefetcher"
version = "0.1.0"
description = "Deterministic price fetching using extraction patterns"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
